requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
DateTime==4.4
decorator==5.1.1
Deprecated==1.2.13
idna==3.3
networkx==2.8.4
numpy==1.22.4
pandas==1.4.2
pathlib==1.0.1
pycparser==2.21
pydiscourse==1.2.0
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.1
ratelimit==2.2.1
requests==2.28.0
six==1.16.0
splitstream==1.2.4
twitter==1.19.6
urllib3==1.26.9
wrapt==1.14.1
xmltodict==0.13.0
zope.interface==5.4.0
